move to infinistore.cpp
